requirements.txt
fastapi>=0.115.0,<1.0
uvicorn>=0.34.0,<1.0
opencv-python>=4.10.0,<5.0
mediapipe>=0.10.14,<1.0
numpy>=1.26.0,<2.0
tensorflow>=2.16.0,<3.0
scikit-learn>=1.5.0,<2.0
joblib>=1.4.0,<2.0
google-generativeai>=0.8.0,<1.0
python-dotenv>=1.0.0,<2.0
Pillow>=10.0.0,<11.0
